User level encryption application.
The script calls the initialization function to check to see if it is in a new directory and then runs a custom shell. This script is meant to be run in a separate and only uses relative file paths. If you would like to set up a  new work space, be sure it is in an empty directory and only run the application in directories which are intended for this purpose.